Rename functional.py to _functional.py
