+++ importlib_resources/_functional.py
